Refactor out element context from GPUI
